minor fix in cluster object saving

+++ src/server/core/cluster.cpp
@@ -315,7 +315,7 @@ bool Cluster::saveToDatabase(DB_HANDLE hdb)
 		}
    }
 
-   if (success)
+   if (success && (m_modified & MODIFY_CLUSTER_RESOURCES))
    {
 		// Save resource list
       hStmt = DBPrepare(hdb, _T("DELETE FROM cluster_resources WHERE cluster_id=?"));
@@ -679,7 +679,7 @@ void Cluster::statusPoll(PollerInfo *poller, ClientSession *pSession, UINT32 dwR
 												 node->getId(), node->getName());
 								}
 								m_pResourceList[k].dwCurrOwner = node->getId();
-								modified |= MODIFIED_CLUSTER_RESOURCES;
+								modified |= MODIFY_CLUSTER_RESOURCES;
 							}
 							resourceFound[k] = 1;
 						}
@@ -708,7 +708,7 @@ void Cluster::statusPoll(PollerInfo *poller, ClientSession *pSession, UINT32 dwR
 							 m_pResourceList[i].dwCurrOwner,
 							 (pObject != NULL) ? pObject->getName() : _T("<unknown>"));
 				m_pResourceList[i].dwCurrOwner = 0;
-            modified |= MODIFIED_CLUSTER_RESOURCES;
+            modified |= MODIFY_CLUSTER_RESOURCES;
 			}
 		}
 		unlockProperties();
